Remove authServer.js and merge code into server.js
